new SDK updates
- moved old SDK to subfolder, pending review for further changes
